pyproject.toml
[project]
name = "plain-mlops-churn"
version = "0.1.0"
description = "Config-driven churn modeling pipeline implemented without Kedro."
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "pandera>=0.17",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "python-dotenv>=1.0",
    "boto3>=1.34",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
